refactor(contracts): split analytics events.ts into per-concern barrel
events.ts (4159 LOC) is now a sub-barrel over ./events/*, one file per
concern: event-names, shared-enums, onboarding, design-systems, page-view,
ui-click, surface-view, result-events, event-payload, mappers, and
design-system-helpers. The public surface (re-exported through the
@open-design/contracts/analytics entry) is byte-identical: 271 exported
names unchanged, no code lines added or removed. Part of #5165.
